requirements.txt
virtualenv
pillow==3.0.0